<feature> Re-structured packages and moved classes to better places <feature> Added Service and CRUDService interfaces to apply Interface Segregation Principle.
