+++ shared/src/main/java/com/makeprojects/ewallet/shared/core/definition/CRUDService.java
@@ -1,0 +1,17 @@
+package com.makeprojects.ewallet.shared.core.definition;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface CRUDService<T> extends Service{
+
+    T get(UUID id);
+
+    List<T> getAll();
+
+    T create(T entity);
+
+    T update(T entity);
+
+    void delete(UUID id);
+}
