+++ shared/src/main/java/com/makeprojects/ewallet/shared/core/definition/Service.java
@@ -1,0 +1,4 @@
+package com.makeprojects.ewallet.shared.core.definition;
+
+public interface Service {
+}
