+++ shared/src/main/java/com/makeprojects/ewallet/shared/database/model/Transaction.java
@@ -1,4 +1,4 @@
-package com.makeprojects.ewallet.shared.model;
+package com.makeprojects.ewallet.shared.database.model;
 
 import com.fasterxml.jackson.annotation.JsonIncludeProperties;
 import jakarta.persistence.*;
