+++ user-accounts/src/main/java/com/makeprojects/ewallet/useraccounts/api/controller/TransactionController.java
@@ -1,13 +1,11 @@
-package com.makeprojects.ewallet.useraccounts.controller;
+package com.makeprojects.ewallet.useraccounts.api.controller;
 
-import com.makeprojects.ewallet.shared.model.Account;
-import com.makeprojects.ewallet.shared.model.Transaction;
+import com.makeprojects.ewallet.shared.database.model.Transaction;
 import com.makeprojects.ewallet.useraccounts.dto.TransactionDto;
-import com.makeprojects.ewallet.useraccounts.service.AccountService;
+import com.makeprojects.ewallet.useraccounts.core.service.implementation.AccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
