+++ user-accounts/src/main/java/com/makeprojects/ewallet/useraccounts/mapper/TransactionMapper.java
@@ -1,11 +1,10 @@
 package com.makeprojects.ewallet.useraccounts.mapper;
 
 import com.makeprojects.ewallet.shared.exceptions.AccountNotFoundException;
-import com.makeprojects.ewallet.shared.model.Account;
-import com.makeprojects.ewallet.shared.model.Transaction;
+import com.makeprojects.ewallet.shared.database.model.Account;
+import com.makeprojects.ewallet.shared.database.model.Transaction;
 import com.makeprojects.ewallet.useraccounts.dto.TransactionDto;
-import com.makeprojects.ewallet.useraccounts.repository.AccountRepository;
-import com.makeprojects.ewallet.useraccounts.service.AccountService;
+import com.makeprojects.ewallet.useraccounts.database.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
